requirements.txt
PyPDF2
python-pptx
requests
tqdm
openrouter
bs4
python-dotenv
pyyaml
